+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 6.2.0 (2024-10-18)
+### Features Added
+
+- New field `StartsAtVersion` in struct `GalleryImageFeature`
+- New field `AllowUpdateImage` in struct `GalleryImageProperties`
+
+
 ## 6.1.0 (2024-08-23)
 ### Features Added